moving-address-fields-to-common-component-and-adding-address-to-department-page
fixed "helpers" folder name to lowercase
