+++ EmployeePortal/ClientApp/src/App.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import { Route, Routes } from 'react-router';
 import { Layout } from './components/Core/Layout';
 import './components/css/Global.css';
-import ComponentWrapper from './components/Helpers/ComponentWrapper';
-import { AuthRoutes } from './components/Helpers/RequireAuth';
+import ComponentWrapper from './components/helpers/ComponentWrapper';
+import { AuthRoutes } from './components/helpers/RequireAuth';
 import { Errors } from './components/Pages/Admin/Errors';
 import { Logs } from './components/Pages/Admin/Logs';
 import { ChangePassword } from './components/Pages/Admin/User/ChangePassword';
