+++ EmployeePortal/ClientApp/src/components/Pages/Admin/Errors.js
@@ -2,7 +2,7 @@ import React from 'react';
 import IsNull, { ReplaceSpecialChars } from '../../common/Common';
 import { DataGrid } from '../../Core/DataGrid';
 import AlertDialog from '../../Core/ModalDialogs';
-import { GetData } from '../../Helpers/WebApi.ts';
+import { GetData } from '../../helpers/WebApi.ts';
 
 export function Errors() {
     let columns, data = []
