+++ EmployeePortal/ClientApp/src/components/Pages/Admin/Logs.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import IsNull from '../../common/Common';
 import { DataGrid } from '../../Core/DataGrid';
-import { useLogService } from './../../Helpers/Logger.ts';
+import { useLogService } from './../../helpers/Logger.ts';
 
 export function Logs() {
     let columns, data = []
