+++ EmployeePortal/ClientApp/src/components/helpers/GetUserInfo.js
