+++ EmployeePortal/ClientApp/src/components/helpers/ImageHelper.js
