+++ EmployeePortal/ClientApp/src/components/helpers/Logger.ts
