+++ EmployeePortal/ClientApp/src/components/helpers/RequireAuth.js
