+++ EmployeePortal/EmployeePortal.csproj
@@ -49,7 +49,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <TypeScriptCompile Remove="ClientApp\src\components\Helpers\Dictionary.ts" />
+    <TypeScriptCompile Remove="ClientApp\src\components\helpers\Dictionary.ts" />
   </ItemGroup>
 
   <Target Name="DebugEnsureNodeEnv" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('$(SpaRoot)node_modules') ">
